refactor(invites): extract HTTP adapter boundary
